Make FD reading portable
It worked on the systems I knew of, but I don't think the casting was
compliant to the standard. The hack I used makes the code compliant I
think, although it introduces a minor special case. It will only enlarge
the struct unnecessarily if int is larger than void*, which is unlikely.