Add small stack optimization
I haven't run any benchmarks, so this might be slower due to the checks.
It also makes the code more complex. Whatever. I think this is a fun
feature. Most stacks do not need to be large. I purposefully kept the
struct size the same by using a union; the ABI is unchanged.